POC: refactor legend
