+++ ui/components/src/Legend/index.ts
@@ -12,3 +12,4 @@
 // limitations under the License.
 
 export * from './Legend';
+export * from './legend-model';
