+++ ui/components/src/model/index.ts
@@ -12,6 +12,5 @@
 // limitations under the License.
 
 export * from './graph';
-export * from './legend';
 export * from './theme';
 export * from './units';
