+++ ui/plugin-system/src/components/LegendOptionsEditor/index.ts
